+++ auth-security/src/test/java/de/alpharogroup/auth/CredentialsTest.java
@@ -25,13 +25,21 @@ import static org.testng.AssertJUnit.assertTrue;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
-import de.alpharogroup.evaluate.object.EqualsEvaluator;
-import de.alpharogroup.evaluate.object.HashcodeEvaluator;
-import de.alpharogroup.evaluate.object.ToStringEvaluator;
+import de.alpharogroup.evaluate.object.evaluators.EqualsEvaluator;
+import de.alpharogroup.evaluate.object.evaluators.HashcodeEvaluator;
+import de.alpharogroup.evaluate.object.evaluators.ToStringEvaluator;
 
+/**
+ * The unit test class for the class {@link Credentials}
+ */
 public class CredentialsTest
 {
 
+
+	/**
+	 * Test method for {@link Credentials#equals(Object)} , {@link Credentials#hashCode()}
+	 * and {@link Credentials#toString()}
+	 */
 	@Test
 	public void testCredentials()
 	{