+++ auth-security/src/test/java/de/alpharogroup/auth/sign/in/BaseUsernameSignInModelTest.java
@@ -24,7 +24,7 @@ import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
 /**
- * The class {@link BaseUsernameSignInModel}.
+ * The unit test class for the class {@link BaseUsernameSignInModel}
  */
 public class BaseUsernameSignInModelTest
 {