+++ auth-security/src/test/java/de/alpharogroup/auth/sign/up/BaseUsernameSignUpModelTest.java
@@ -24,7 +24,7 @@ import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
 /**
- * The class {@link BaseUsernameSignUpModel}.
+ * The unit test class for the class {@link BaseUsernameSignUpModel}
  */
 public class BaseUsernameSignUpModelTest
 {