+++ crypt-data/src/test/java/de/alpharogroup/crypto/obfuscation/rule/ObfuscationOperationRuleTest.java
@@ -26,7 +26,7 @@ import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
 import de.alpharogroup.collections.set.SetFactory;
-import de.alpharogroup.evaluate.object.EqualsHashCodeAndToStringEvaluator;
+import de.alpharogroup.evaluate.object.evaluators.EqualsHashCodeAndToStringEvaluator;
 
 /**
  * The unit test class for the class {@link ObfuscationOperationRule}