+++ crypt-data/src/test/java/de/alpharogroup/crypto/obfuscation/rules/SimpleObfuscationRulesTest.java
@@ -33,7 +33,7 @@ import com.google.common.collect.BiMap;
 import com.google.common.collect.HashBiMap;
 
 import de.alpharogroup.collections.map.MapFactory;
-import de.alpharogroup.evaluate.object.EqualsHashCodeAndToStringEvaluator;
+import de.alpharogroup.evaluate.object.evaluators.EqualsHashCodeAndToStringEvaluator;
 
 /**
  * The unit test class for the class {@link ObfuscationBiMapRules}.